backend/requirements.txt
flask
flask-cors
flask-sqlalchemy
pyjwt
bcrypt
gunicorn
